refactor(ios/engine): KMPMetadata assertions on package extraction tests
