+++ ios/engine/KMEI/KeymanEngineTests/FileManagementTests.swift
@@ -21,56 +21,6 @@ class FileManagementTests: XCTestCase {
     TestUtils.standardTearDown()
   }
 
-  func testKeyboardPackageExtraction() throws {
-    let cacheDirectory = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask)[0]
-    let khmerPackageZip = cacheDirectory.appendingPathComponent("khmer_angkor.zip")
-    try FileManager.default.copyItem(at: TestUtils.Keyboards.khmerAngkorKMP, to: khmerPackageZip)
-
-    let destinationFolderURL = cacheDirectory.appendingPathComponent("khmer_angkor")
-
-    // Requires that the source file is already .zip, not .kmp.  It's a ZipUtils limitation.
-    do {
-      try KeymanPackage.extract(fileUrl: khmerPackageZip, destination: destinationFolderURL, complete: { kmp in
-        if let kmp = kmp {
-          XCTAssertTrue(kmp.isKeyboard(), "Keyboard KMP test extraction did not yield a keyboard package!")
-
-          // extracted ok, test kmp
-          XCTAssert(kmp.sourceFolder == destinationFolderURL,
-                    "The KMP's reported 'source folder' should match the specified destination folder")
-        } else {
-          XCTAssert(false, "KeymanPackage.extract failed")
-        }
-      })
-    } catch {
-      XCTFail("KeymanPackage.extract failed with error \(error)")
-    }
-  }
-
-  func testLexicalModelPackageExtraction() throws {
-    let cacheDirectory = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask)[0]
-    let mtntZip = cacheDirectory.appendingPathComponent("mtnt.zip")
-    try FileManager.default.copyItem(at: TestUtils.LexicalModels.mtntKMP, to: mtntZip)
-
-    let destinationFolderURL = cacheDirectory.appendingPathComponent("mtnt.model")
-
-    // Requires that the source file is already .zip, not .kmp.  It's a ZipUtils limitation.
-    do {
-      try KeymanPackage.extract(fileUrl: mtntZip, destination: destinationFolderURL, complete: { kmp in
-        if let kmp = kmp {
-          XCTAssertTrue(!kmp.isKeyboard(), "Lexical model KMP test extraction yielded a keyboard package!")
-
-          // extracted ok, test kmp
-          XCTAssert(kmp.sourceFolder == destinationFolderURL,
-                    "The KMP's reported 'source folder' should match the specified destination folder")
-        } else {
-          XCTAssert(false, "KeymanPackage.extract failed")
-        }
-      })
-    } catch {
-      XCTFail("KeymanPackage.extract failed with error \(error)")
-    }
-  }
-
   func testKeyboardInstallation() {
     ResourceFileManager.shared.prepareKMPInstall(from: TestUtils.Keyboards.khmerAngkorKMP) { kmp, error in
       XCTAssertNotNil(kmp, "Failed to prepare KMP for installation")
